Tweak the widget and render API
The render function now looks like:

	Renderer.render(tiddler, context, template);

with template being optional. This means you can now pass a string in as
the tiddler if you've attached a store to the renderer.

The widget syntax has changed the handler function so that the args are
now passed as an argument to the function. The context has also changed,
so the handler is now called like this:

	handler.call({ store: store, args: args}, args);